remove redundant cast

+++ vuzeAndroidRemote/src/main/java/com/vuze/android/remote/ActionBarToolbarSplitter.java
@@ -103,12 +103,12 @@ public class ActionBarToolbarSplitter
 
 		widthRemaining -= widthPadding;
 
-		int hardCodedPaddingPx = (int) AndroidUtilsUI.dpToPx(GENERATED_ITEM_PADDING);
+		int hardCodedPaddingPx = AndroidUtilsUI.dpToPx(GENERATED_ITEM_PADDING);
 
 		int padding = hardCodedPaddingPx * 2;
 
 
-		int minIconWidth = (int) AndroidUtilsUI.dpToPx(MIN_CELL_SIZE);
+		int minIconWidth = AndroidUtilsUI.dpToPx(MIN_CELL_SIZE);
 
 		int hardCodedOverFlowWidth = minIconWidth;
